Move more classes and code out of Resolver.cs
